project/requirements.txt
Django==3.0.5
django_extensions==2.2.9
djangorestframework==3.11.0
django-model-utils==4.0.0
requests==2.23.0
paypalrestsdk==1.13.1
pillow==7.1.1
django-cors-headers==3.2.1
django-constance==2.6.0
pycrypto==2.6.1
reportlab==3.5.42
coreapi==2.3.3
iPython==7.13.0
django-jet==1.0.8
stripe==2.45.0
#Microsoft Visual C++ 9.0 is required. Get it from http://aka.ms/vcpython27